requirement.txt
langchain-core==0.3.68
langgraph==0.5.2
langchain-community==0.3.27
langchain-openai==0.3.27
python-dotenv==1.1.1
PyYAML==6.0.2
#pysqlite3-binary
pydantic==2.11.7
dashscope==1.23.8
google-api-python-client==2.176.0
gradio==5.38.0
pillow==10.4.0
google-search-results==2.4.2

#Python 3.12.7